pyproject.toml
[tool.poetry]
name = "instloader"
version = "0.1.0"
description = ""
authors = ["aorews <aorew@yandex.ru>"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.8"
pytest = "^7.2.2"
black = "^23.1.0"
coverage = "^7.2.1"
python-dotenv = "^1.0.0"
instaloader = "^4.9.6"


[tool.poetry.group.dev.dependencies]
black = "^23.1.0"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
